create_iteration | improvements

+++ tdeptools/cli/tdep_create_next_iteration.py
@@ -11,7 +11,36 @@ from .tdep_create_sample_folders import main as create_sample_folders
 
 
 _cmd = "canonical_configuration"
-_infiles = ("infile.ucposcar", "infile.ssposcar", "infile.forceconstant")
+_infiles = (
+    "infile.ucposcar",
+    "infile.ssposcar",
+    "infile.forceconstant",
+    "infile.lotosplitting",
+)
+
+
+def _create_samples(
+    temperature: float,
+    nsamples: int,
+    mf: float,
+    quantum: bool,
+    folder_new: Path,
+    of: int = 4,
+):
+    cmd = f"{_cmd} -t {temperature} -of 4 -n {nsamples}"
+
+    if mf is not None:
+        cmd += " -mf {mf}"
+
+    if quantum:
+        cmd += " --quantum"
+
+    ps = sp.run(cmd.split(), cwd=folder_new, capture_output=True)
+    stdout = ps.stdout + ps.stderr
+    with open(folder_new / (_cmd + ".log"), "wb") as f:
+        f.write(cmd.encode())
+        f.write("\n".encode())
+        f.write(stdout)
 
 
 app = typer.Typer()
@@ -22,6 +51,7 @@ def main(
     prefix: str = "iter.",
     temperature: float = typer.Option(None, "--temperature", "-T"),
     mf: float = None,
+    of: int = 4,
     quantum: bool = True,
     force: bool = False,
     makefile: str = "Makefile",
@@ -47,22 +77,17 @@ def main(
     folder_new.mkdir(exist_ok=force)
 
     for file in _infiles:
-        shutil.copy(file, folder_new)
+        if Path(file).exists():
+            shutil.copy(file, folder_new)
 
-    cmd = f"{_cmd} -t {temperature} -of 4 -n {nsamples}"
-
-    if mf is not None:
-        cmd += " -mf {mf}"
-
-    if quantum:
-        cmd += " --quantum"
-
-    ps = sp.run(cmd.split(), cwd=folder_new, capture_output=True)
-    stdout = ps.stdout + ps.stderr
-    with open(folder_new / (_cmd + ".log"), "wb") as f:
-        f.write(cmd.encode())
-        f.write("\n".encode())
-        f.write(stdout)
+    _create_samples(
+        temperature=temperature,
+        nsamples=nsamples,
+        mf=mf,
+        quantum=quantum,
+        folder_new=folder_new,
+        of=of,
+    )
 
     # copy files
     echo(".. copy Makefile and control")
